# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "api/comment-reply": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "User Create Comments Reply",
                "description": "User Create Comments Reply",
                "operationId": "User Create Comments Reply",
                "requestBody": {
                    "description": "User Create Comments",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "true"
                                ],
                                "properties": {
                                    "comment_reply": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "some comment"
                                    },
                                    "comment_id": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User Create Comments",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "api/replyanswer_viewed-notification": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "User viewed Reply Answer Comments",
                "description": "User viewed Reply Answer Comments",
                "operationId": "Users viewed Reply Answer Comments",
                "requestBody": {
                    "description": "Users viewed Reply Answer Comments",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "true"
                                ],
                                "properties": {
                                    "notification_id": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User viewed Reply Answer Comments",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "api/comment-reply-answer": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "User Create Comments Replys",
                "description": "User Create Comments Replys",
                "operationId": "User Create Comments Replys",
                "requestBody": {
                    "description": "User Create Comments Reply",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "true"
                                ],
                                "properties": {
                                    "reply": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "some comment"
                                    },
                                    "reply_id": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User Create Comments",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    }
}